# 8278302: [s390] Implement fast-path for ASCII-compatible CharsetEncoders