This pull request contains the s390x variant of the fast-path for ASCII-compatible CharsetEncoders. 

The intrinsic implementation is pretty much similar to that of PPC64. Compared to the previously existing intrinsic implementation, changes to program logic are minimal. There is a new match rule for ascii strings and the bit mask used to detect invalid bits in the string characters is set depending on the target charset: 0xff00 for ISO, 0xff80 for ascii.

**Request:** SAP does not maintain a full build and test infrastructure for s390x anymore. Could somebody please test this PR on s390x? At least build and tier1 tests, the more the better. Thank you!

With my private testing (some microbenchmarks, some SPEC benchmarks) I could not find any issues. With the fast-path active I see a 2x, sometime 2.5x, improvement of CharsetEncodeDecode.encode. 

**Without the fast-path patch:**
```
CharsetEncodeDecode.encode   16384        UTF-8  avgt   30  16.350 ± 0.067  us/op
CharsetEncodeDecode.encode   16384         BIG5  avgt   30  15.972 ± 0.136  us/op
CharsetEncodeDecode.encode   16384  ISO-8859-15  avgt   30  14.372 ± 0.110  us/op
CharsetEncodeDecode.encode   16384        ASCII  avgt   30  14.287 ± 0.051  us/op
```

**With the fast-path patch:**
```
CharsetEncodeDecode.encode   16384        UTF-8  avgt   30   6.833 ± 0.064  us/op
CharsetEncodeDecode.encode   16384         BIG5  avgt   30   8.195 ± 0.059  us/op
CharsetEncodeDecode.encode   16384  ISO-8859-15  avgt   30   6.387 ± 0.149  us/op
CharsetEncodeDecode.encode   16384        ASCII  avgt   30   6.321 ± 0.049  us/op
```

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue
- [x] Change must be properly reviewed

### Issue
 * [JDK-8278302](https://bugs.openjdk.java.net/browse/JDK-8278302): [s390] Implement fast-path for ASCII-compatible CharsetEncoders


### Reviewers
 * [Martin Doerr](https://openjdk.java.net/census#mdoerr) (@TheRealMDoerr - **Reviewer**) ⚠️ Review applies to 4326bf93de4dd76edd78cee044e2a2b44a2822d2
 * @backwaterred (no known github.com user name / role)


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.java.net/jdk pull/6738/head:pull/6738` \
`$ git checkout pull/6738`

Update a local copy of the PR: \
`$ git checkout pull/6738` \
`$ git pull https://git.openjdk.java.net/jdk pull/6738/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 6738`

View PR using the GUI difftool: \
`$ git pr show -t 6738`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.java.net/jdk/pull/6738.diff">https://git.openjdk.java.net/jdk/pull/6738.diff</a>

</details>