@RealLucy 2022-01-04
Had already contacted Tim Ellison (RedHat) before Christmas. He promised to point an engineer to this PR.