@backwaterred 2022-01-10
@RealLucy, sorry to keep you waiting. I am currently running tier1-tests on our s390x system. I'll edit this comment with the results when it completes.

---