@RealLucy 2022-01-10
@backwaterred Thanks for working on the tests.
At first glance, the failure you are seeing doesn't immediately suggest it's my changes that are causing it. Do I understand the text "INTERPRETED" correctly if I assume the test was running without JIT compilation?