@backwaterred 2022-01-11
I am not very familiar with it, but taking a quick look at [the test](https://github.com/openjdk/jdk/blob/master/test/hotspot/jtreg/compiler/unsafe/UnsafeCopyMemory.java) I think you are right. Since the failure occurred after only a few tests in the run, I believe the failure occurred in the interpreted section.