@RealLucy 2022-01-11
Thank you for testing! The results don't look good, though. I'll have to find out what the test expects and why my code doesn't deliver the expected. Will get back to you asap. 
Lutz