@RealLucy 2022-01-17
@backwaterred I have updated the PR with a commit fixing the issue found during testing. The only code change in this commit is in line #80, where I removed the "int", turning the declaration/definition into an assignment. All else is just code comments. 

I would appreciate very much if you could run the tests once again. There is one more thing: could you, or somebody else in your team, please review the change? Thx.