@RealLucy 2022-01-17
@backwaterred Yes, for hotspot changes, two reviews are required. At least one review must be from a "real" reviewer.