@RealLucy 2022-01-17
Let me aid understanding with a few additional comments. 

The implementation of C2_MacroAssembler::string_compress() was pre-existing for the plain compress function and for encoding ISO (8-bit character) arrays. What I had to add was another mask value to support 7-bit ascii characters. The encoding logic remained the same. 

In the s390.ad file, a new match rule was added for 7-bit ascii characters. This rule is, with the exception of the predicate, identical to the one for ISO arrays.