@RealLucy 2022-01-18
Thank you for your comments, Claes!
As SAP has passed on responsibility for the s390 port, I would like to receive a "lead" review from the new maintainers. Then, I could recruit a "second opinion" from my direct colleagues.